#149 반환 값 없는 조회 메서드 제거 및 boolean 타입 메서드로 변경

+++ src/main/java/com/example/gatherplan/appointment/service/impl/UserServiceImpl.java
@@ -179,8 +179,9 @@ public class UserServiceImpl implements UserService, UserDetailsService {
     @Override
     @Transactional
     public void authenticateEmailForPasswordReset(String email) {
-        userRepository.findByEmail(email).orElseThrow(() ->
-                new UserException(ErrorCode.USER_NOT_FOUND));
+        if (!userRepository.existsByEmail(email)) {
+            throw new UserException(ErrorCode.USER_NOT_FOUND);
+        }
 
         emailAuthRepository.findByEmail(email)
                 .ifPresent(emailAuth -> emailAuthRepository.deleteByEmail(emailAuth.getEmail()));
